+++ kurtosis-module/impl/participant_network/el/besu/besu_launcher.go
@@ -88,8 +88,6 @@ func (launcher *BesuELClientLauncher) Launch(
 
 	miningWaiter := mining_waiter.NewMiningWaiter(restClient)
 	result := el.NewELClientContext(
-		// TODO Figure out how to get the ENR so CL clients can connect to it!!
-		"", // Besu node info endpoint doesn't return an ENR
 		nodeInfo.Enode,
 		serviceCtx.GetPrivateIPAddress(),
 		rpcPortNum,