+++ kurtosis-module/impl/participant_network/el/el_rest_client/api_response_objects.go
@@ -19,5 +19,4 @@ type GetNodeInfoResponse struct {
 
 type NodeInfo struct {
 	Enode string	`json:"enode"`
-	ENR   string	`json:"enr"`
 }