+++ kurtosis-module/impl/participant_network/el/geth/geth_launcher.go
@@ -103,7 +103,6 @@ func (launcher *GethELClientLauncher) Launch(
 
 	miningWaiter := mining_waiter.NewMiningWaiter(restClient)
 	result := el.NewELClientContext(
-		nodeInfo.ENR,
 		nodeInfo.Enode,
 		serviceCtx.GetPrivateIPAddress(),
 		rpcPortNum,