+++ kurtosis-module/impl/participant_network/el/nethermind/nethermind_launcher.go
@@ -83,8 +83,6 @@ func (launcher *NethermindELClientLauncher) Launch(
 
 	miningWaiter := mining_waiter.NewMiningWaiter(restClient)
 	result := el.NewELClientContext(
-		// TODO TODO TODO TODO Get Nethermind ENR, so that CL clients can connect to it!!!
-		"", //Nethermind node info endpoint doesn't return ENR field https://docs.nethermind.io/nethermind/ethereum-client/json-rpc/admin
 		nodeInfo.Enode,
 		serviceCtx.GetPrivateIPAddress(),
 		rpcPortNum,